requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.4
bokeh==2.4.3
boto3==1.24.89
botocore==1.27.89
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
cryptography==37.0.4
cycler==0.11.0
dm-tree==0.1.7
filelock==3.9.0
flatbuffers==1.12
fonttools==4.37.1
frozenlist==1.3.1
gast==0.4.0
gdown==4.6.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.4
h5py==3.7.0
humanfriendly==10.0
idna==3.3
imageio==2.21.2
imgaug==0.4.0
importlib-metadata==4.11.4
Jinja2==3.1.2
jmespath==1.0.1
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mpmath==1.2.1
multidict==6.0.2
networkx==2.8.6
numpy==1.23.2
oauthlib==3.2.0
onnxruntime==1.12.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pip==22.2.2
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.4
pyu2f==0.1.5
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-image==0.19.3
scipy==1.9.0
setuptools==65.3.0
Shapely==1.8.4
six==1.16.0
soupsieve==2.3.2.post1
sympy==1.11
tensorboard==2.9.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-model-optimization==0.7.3
termcolor==1.1.0
tifffile==2022.8.12
tornado==6.2
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.1
gradio==3.28.0
